reorganized a few files
