--- src/main/java/app/controllers/MenuBarController.java
@@ -1,14 +1,0 @@
-package app.controllers;
-
-import app.interaction_managers.Destroyer;
-import app.interaction_managers.Zoomer;
-
-public class MenuBarController {
-    public void zoom() {
-        Zoomer zoomer = new Zoomer();
-    }
-
-    public void destroy() {
-        Destroyer destroyer = new Destroyer();
-    }
-}
