--- src/main/java/app/interaction_managers/Destroyer.java
@@ -1,21 +1,0 @@
-package app.interaction_managers;
-
-import app.interaction_managers.InteractionManager;
-
-public class Destroyer implements InteractionManager {
-
-    private Object toRemove;
-
-    @Override
-    public void interact() {
-
-    }
-
-    public void removeFromPage(Object toRemove){
-
-    }
-
-    public void removeRefrences (Object toRemove){
-
-    }
-}
