--- src/main/java/app/interaction_managers/Zoomer.java
@@ -1,17 +1,0 @@
-package app.interaction_managers;
-
-import app.interaction_managers.InteractionManager;
-
-public class Zoomer implements InteractionManager {
-
-    private Object toZoom;
-
-    @Override
-    public void interact() {
-
-    }
-
-    public void zoom(Object toZoom){
-
-    }
-}
